Refactor component imports and file structure BugFix: Verify Token was issuing Tokens when not logged, Restructured to only verify if a token is already set.
