+++ src/views/Logout.jsx
@@ -1,11 +1,15 @@
 import { useEffect } from 'react';
 import { useAuth } from '../hooks/useAuthProvider';
 
+import { useNavigate } from 'react-router-dom';
+
 const Logout = () => {
+    const navigate = useNavigate();
     const { logout } = useAuth();
 
     useEffect(() => {
         logout();
+        navigate("/login");
     }, []); // eslint-disable-line react-hooks/exhaustive-deps
